Remove compare from the array prototype
Overloading a built in classes prototype is not a terribly good idea;
unless you are doing it for cross-browser compatibility (i.e. IE8
doesn't have map; so you provide an implementation of map on the Array
prototype in IE 8 only)

There's a notion in design for "Open for Extension but closed for
Modification" when you add an additional function to a prototype you're
modifying that types behavior. In the event someone *else* depends on
your compare function and then you change it (perhaps it also supports
hashmaps!) you are setting them up for pain.