Add some comments for future direction of the move function
Currently every single one of your east/north forwards/backwards
functions have to check to see if you leave the world or if you run into
an obstacle.

Also, it would be better to have a single datatype for coordinate,
instead of going back and forth between the array and the coordinate
with x/y etc.

Your parameters object is actually several objects at once: a
coordinate, a set of commands, a set of obstacles, etc. Instead of
passing all of this data to every function, think about how you could
only pass the data the given function needs